fix: use float provider revenue for solver, aligning with reported metrics
provider_revenue (p_ij) is a monetary value and may be fractional. The unit
aggregation in _build_payload summed int(provider_revenue) per member, and the
PyO3 boundary cast the p matrix to int64, so the solver saw a truncated p in
constraint (4) and F_prov — while compute_provider_metrics reports the raw
float p. For service groups the per-member truncation accumulated, biasing
constraint (4) stricter than reported and optimizing F_prov against an
understated value.

De-truncate p end-to-end: sum raw float in _build_payload, type p_ij as float
in the Pydantic inputs, cast to np.float64 at the boundary, and store p_ij as
Array2<f64> in the Rust CombinedTask (is_admissible / provider_objective /
provider_heuristic now take f64). Price and resource stay int (article §5.1) —
they are consistent between solver and metrics already. The weighted-discount
omega aggregation, s_unit=max, and one-discount-per-group are unchanged.

No DB migration: provider_revenue is Numeric(18,4); provider_value /
provider_profit are double precision.

+++ apps/worker/src/worker/activities.py
@@ -50,9 +50,9 @@ async def _run_algorithm_core(input: RunAlgorithmInput) -> RunResult:
     # s = 0 → every pair admissible (identical to the pre-(4) behaviour).
     m, n = input.m, input.n
     p_ij = (
-        np.array(input.p_ij, dtype=np.int64)
+        np.array(input.p_ij, dtype=np.float64)
         if input.p_ij is not None
-        else np.zeros((m, n), dtype=np.int64)
+        else np.zeros((m, n), dtype=np.float64)
     )
     s_ij = (
         np.array(input.s_ij, dtype=np.float64)
@@ -163,7 +163,7 @@ async def run_combined_method_activity(input: CombinedSolveInput) -> CombinedRes
         input.n,
         np.array(input.c, dtype=np.int64),
         np.array(input.b_ij, dtype=np.int64),
-        np.array(input.p_ij, dtype=np.int64),
+        np.array(input.p_ij, dtype=np.float64),
         np.array(input.omega_max, dtype=np.float64),
         np.array(input.s_ij, dtype=np.float64),
         int(input.b_total),
